9/10 V2.0 change BigDecimal, add @Valid

+++ esmp_demo/src/main/java/com/practice/esmp_demo/controller/dto/UpdateMstmb.java
@@ -4,11 +4,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 public class UpdateMstmb {
-    private double curPrice;
+    private BigDecimal curPrice;
     private String modDate;
     private String modTime;
     private String modUser;
